+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - Removed route selection for IDS subscription
 - Breadcrumbs: remove null links
 - Docker Image with Non-Root-User
+- Delete representation & artifact when requested resource is deleted
 
 ## [9.2.0] - 2022-04-08 (compatible with DSC 7.0.0)